Scheduling Refactoring and docs improvement
