+++ geco/mips/tests/test_scheduling.py
@@ -1,19 +1,18 @@
 import pytest
 
-from geco.mips.scheduling import *
+from geco.mips.scheduling.heinz import *
+from geco.mips.scheduling.hooker import *
 
 
-def test_late_tasks_formulation():
-    main_params = number_of_facilities, number_of_tasks, time_steps = 3, 10, 100
-    params = p, C, c, R, d = generate_params(*main_params[:-1], seed=0)[:-1]
-
-    model = hooker_late_tasks_formulation(*main_params, *params)
-    check_hookers_instance(model, number_of_facilities, number_of_tasks, time_steps)
+def test_hooker_formulation():
+    params = number_of_facilities, number_of_tasks, time_steps = 3, 10, 100
+    model = hooker_instance(*params)
+    _check_hookers_instance(model, number_of_facilities, number_of_tasks, time_steps)
 
 
 def test_hooker_generation():
     for params, model in generate_hookers_instances():
-        check_hookers_instance(model, *params[:-1])
+        _check_hookers_instance(model, *params[:-1])
 
 
 @pytest.mark.parametrize(
@@ -22,9 +21,9 @@ def test_hooker_generation():
 )
 def test_heinz_formulation(number_of_facilities, number_of_tasks, seed):
     main_params = number_of_facilities, number_of_tasks
-    params = p, C, c, R, d, r = generate_params(*main_params, seed)
+    p, C, c, R, d, r = heinz_params(*main_params, seed)
     time_steps = int(max(d.values()) - min(R))
-    model = heinz_formulation(*main_params, *params)
+    model = heinz_instance(*main_params)
     x_vars_count = number_of_facilities * number_of_tasks
     y_vars_lowerbound = 0
     y_vars_upperbound = number_of_facilities * number_of_tasks * time_steps
@@ -68,7 +67,7 @@ def test_param_generation_seeding(n_resources, n_tasks, seed1, seed2):
 
 
 def test_hooker_simple_instance():
-    hooker_model = hooker_late_tasks_formulation(*simple_instance_params()[:-1])
+    hooker_model = hooker_formulation(*_simple_instance_params()[:-1])
     hooker_model.hideOutput()
     hooker_model.optimize()
     assert hooker_model.getStatus() == "optimal"
@@ -86,7 +85,7 @@ def test_heinz_simple_instance():
         release_times,
         deadlines,
         resource_requirements,
-    ) = simple_instance_params()
+    ) = _simple_instance_params()
     heinz_model = heinz_formulation(
         n_resources,
         n_tasks,
@@ -103,7 +102,7 @@ def test_heinz_simple_instance():
     assert heinz_model.getObjVal() == 1
 
 
-def simple_instance_params():
+def _simple_instance_params():
     n_resources = 1
     n_tasks = 1
     time_steps = 1
@@ -128,7 +127,7 @@ def simple_instance_params():
     )
 
 
-def check_hookers_instance(model, number_of_facilities, number_of_tasks, time_steps):
+def _check_hookers_instance(model, number_of_facilities, number_of_tasks, time_steps):
     assert (
         model.getNVars()
         == number_of_facilities * number_of_tasks * time_steps + number_of_tasks
